refactor: use mixins instead of base classes
