+++ addon/components/dt-sortable-column-header.js
@@ -1,6 +1,7 @@
-import SortableColumn from './dt-sortable-column-base';
+import Component from '@ember/component';
+import SortableColumn from '../mixins/sortable-column';
 
-export default SortableColumn.extend({
+export default Component.extend(SortableColumn, {
   tagName:'th',
   header: true,
   classNames:['contextual-sortable-header-cell']
