+++ addon/mixins/filterable-column.js
@@ -1,11 +1,11 @@
-import Component from '@ember/component';
+import Mixin from '@ember/object/mixin';
 import { observer } from '@ember/object';
-import layout from '../templates/dt-filterable-column-base';
+import layout from '../templates/dt-filterable-column';
 
-export default Component.extend({
+export default Mixin.create({
   layout,
 
-  filterValueChanged:observer('filterValue', function (){
+  filterValueChanged: observer('filterValue', function (){
     let actionHandler = this.get('filterinformationupdated') || function(){};
     actionHandler(this.get('propertyName'), this.get('filterValue'));
   }),
