+++ addon/templates/dt-filterable-column.hbs
